% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{checktable}
\alias{checktable}
\title{Check if names of the setup table are subset of the count matrix column names}
\usage{
checktable(alleleSpecific = FALSE)
}
\arguments{
\item{alleleSpecific}{TRUE/FALSE is the design allele-specific?}
}
\value{
NA
}
\description{
Check if names of the setup table are subset of the count matrix column names
}
\examples{
checktable(FALSE)

}

