% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{DESeq_allelic}
\alias{DESeq_allelic}
\title{DESeq using allele-specific design}
\usage{
DESeq_allelic(countdata, coldata, fdr)
}
\arguments{
\item{countdata}{count file from featurecounts}

\item{coldata}{sampleInfo file}

\item{fdr}{fdr cutoff for DE}
}
\value{
A list of dds and ddr objects
}
\description{
DESeq using allele-specific design
}
\examples{
DESeq_allelic(countdata, coldata = sampleInfo, fdr = 0.05)

}

