% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{DESeq_downstream}
\alias{DESeq_downstream}
\title{Plotting and annotation of DESeq outputs}
\usage{
DESeq_downstream(DEseqout, countdata, coldata, fdr, outprefix, heatmap_topN,
  geneNamesFile)
}
\arguments{
\item{DEseqout}{output from DEseq_basic or DEseq_allelic wrapper}

\item{countdata}{count file from featurecounts}

\item{coldata}{sampleInfo file}

\item{fdr}{fdr cutoff for DE}

\item{outprefix}{prefix of output plots and files}

\item{heatmap_topN}{top N genes to plot in the heatmap}

\item{geneNamesFile}{tsv file with gene symbols corresponding to ensembl gene IDs}
}
\value{
DEseq results (.tsv) and plots (.pdf)
}
\description{
Plotting and annotation of DESeq outputs
}
\examples{
DESeq_downstream(DEseqout = seqout, countdata, sampleInfo,
fdr = fdr, outprefix = "DEseq_basic", heatmap_topN = topN,
geneNamesFile = geneNamesFilePath)

}

