% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_functions.R
\name{filterByGlobal_chip}
\alias{filterByGlobal_chip}
\title{Filtering using global bg}
\usage{
filterByGlobal_chip(chipCountObject)
}
\arguments{
\item{chipCountObject}{output from readfiles_chip}
}
\value{
Filtered chipCountObject
}
\description{
Filtering using global bg
}
\examples{
filterByInput_chip(chipCountObject,priorCount = 5)

}

