% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_functions.R
\name{filterByInput_chip}
\alias{filterByInput_chip}
\title{Filtering using Input windows}
\usage{
filterByInput_chip(chipCountObject, controlbam, chipbam, priorCount = 5)
}
\arguments{
\item{chipCountObject}{output from readfiles_chip}

\item{controlbam}{control bam file}

\item{chipbam}{chip bam file}

\item{priorCount}{Minimum count cutoff for windows}
}
\value{
Filtered chipCountObject
}
\description{
Filtering using Input windows
}
\examples{
filterByInput_chip(chipCountObject,priorCount = 5)

}

