% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_functions.R
\name{getDBregions_chip}
\alias{getDBregions_chip}
\title{Test for Diff Bound windows using EdgeR (then merge windows into regions)}
\usage{
getDBregions_chip(chipCountObject, plotfile = NULL, tfname)
}
\arguments{
\item{chipCountObject}{output from tmmNormalize_chip}

\item{plotfile}{file with output plots}

\item{tfname}{which TF to extract results for (must match with the name in sampleInfo)}
}
\value{
chipResultObject with differentially bound regions
}
\description{
Test for Diff Bound windows using EdgeR (then merge windows into regions)
}
\examples{
getDBregions_chip(chipCountObject,plotfile = NULL, tfname = "msl2")

}

