% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_functions.R
\name{makeQCplots_chip}
\alias{makeQCplots_chip}
\title{Make plots to select window size and pe-distance cutoffs}
\usage{
makeQCplots_chip(bam.file, outplot, pe.param)
}
\arguments{
\item{bam.file}{bam file to use}

\item{outplot}{path to output pdf}

\item{pe.param}{parameters to read bam files}
}
\value{
QC plots as pdf
}
\description{
Make plots to select window size and pe-distance cutoffs
}
\examples{
makeQCplots_chip(bam.file, outplot)

}

