% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_functions.R
\name{readfiles_chip}
\alias{readfiles_chip}
\title{Read the Files and Count windows for ChIP-Seq Samples}
\usage{
readfiles_chip(sampleInfo, fragment_length, window_size,
  alleleSpecific = FALSE, pe.param)
}
\arguments{
\item{sampleInfo}{tsvfile with sample information}

\item{fragment_length}{fragment length of sequencing}

\item{window_size}{window size to count the reads in}

\item{alleleSpecific}{TRUE/FALSE whether samples have to be compared with allele-specific design}

\item{pe.param}{parameters to read bam files}
}
\value{
chipCountObject : a list with window counts and sampleinfo
}
\description{
Read the Files and Count windows for ChIP-Seq Samples
}
\examples{
readfiles_chip(csvFile = "testBAMs/testSampleSheet.csv", refAllele = "pat")

}

