% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_functions.R
\name{tmmNormalize_chip}
\alias{tmmNormalize_chip}
\title{TMM normalize (get the normfactors out) using given window size}
\usage{
tmmNormalize_chip(chipCountObject, binsize, plotfile)
}
\arguments{
\item{chipCountObject}{output from readfiles_chip command}

\item{binsize}{Size of bins to calculate the normalization factors}

\item{plotfile}{file with output plots}
}
\value{
Normalized chipCountObject
}
\description{
TMM normalize (get the normfactors out) using given window size
}
\examples{
tmmNormalize_chip(chipCountObject,binsize = 10000, plotfile = "TMM_normalizedCounts.pdf")

}

