% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_functions.R
\name{writeOutput_chip}
\alias{writeOutput_chip}
\title{Annotate and print the output regions}
\usage{
writeOutput_chip(chipResultObject, outfile_prefix, fdrcutoff)
}
\arguments{
\item{chipResultObject}{output from getDBregions_chip}

\item{outfile_prefix}{name of output files}

\item{fdrcutoff}{fdr cutoff to call significantly bound regions}
}
\value{
File with differentially bound regions
}
\description{
Annotate and print the output regions
}
\examples{
writeOutput_chip(chipResultObject, outfile_prefix)

}

